boxes iteration: Change the for loop to a foreach
The for loop was assuming that the user was using array keys
sequentially in the boxes array.  While this would normally be the case,
the example code encourages you to skip a key as the example for
'Presence Server' used a commented out key of '1' - the symptom here was
that the mi_connectors array was not built correctly.