[dialplan] fix locating attributes with number values.
If the array of attributes contains numberical keys like, the strpos fails. According to docs -> if needle is not a string, it is converted to an integer and applied as the ordinal value of a character.
Reported by Flavio Goncalves.
Example of failing configuration:

$config->attrs_cb=array(
                 // name , description
                "3" => "usrloc",
                "4" => "pstn",
                "5" => "media",
         );
Note the numerical keys!

(cherry picked from commit c4028b683ba117fd24948b3d7920fdbf373ceaed)